+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fix migration error if http method is missing ([#12](https://github.com/scm-manager/scm-webhook-plugin/pull/12))
+
 ## [2.1.0] - 2020-07-23
 ### Added
 - Documentation in German ([#5](https://github.com/scm-manager/scm-webhook-plugin/pull/5))