+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.2 - 2022-11-28
+### Fixed
+- Error on webhook modifications in repositories ([#53](https://github.com/scm-manager/scm-webhook-plugin/pull/53))
+
 ## 2.3.1 - 2022-11-24
 ### Fixed
 - Exception for new configuration ([#52](https://github.com/scm-manager/scm-webhook-plugin/pull/52))